Embassy Synchronizer
The `embassy-sync` crate provides abstractions for task communication,
similar to the `std::sync` crate. This program now runs three tasks
concurrently. Two for each of the buttons where when a button is pressed
it signals the statically created `Signal` struct. One for reading
from the temperature sensor and wait for incoming signals and adjust
the reading delay time.